+++ src/ts/src/public/time-entry/TimeEntryCreate.ts
@@ -1,0 +1,11 @@
+import { TimeEntryDto } from "./dto";
+
+export interface TimeEntryCreateRequest {
+  start: Date;
+  end: Date;
+  description: string;
+}
+
+export interface TimeEntryCreateResponse {
+  item: TimeEntryDto;
+}