+++ src/ts/src/public/time-entry/TimeEntryGet.ts
@@ -1,0 +1,5 @@
+import { TimeEntryDto } from "./dto";
+
+export interface TimeEntryGetResponse {
+  item: TimeEntryDto;
+}