+++ src/ts/src/public/time-entry/TimeEntryUpdate.ts
@@ -1,0 +1,10 @@
+import { TimeEntryDto, TimeEntryUpdateValues } from "./dto";
+
+export interface TimeEntryUpdateRequest {
+  id: string;
+  values: TimeEntryUpdateValues;
+}
+
+export interface TimeEntryUpdateResponse {
+  item: TimeEntryDto;
+}