+++ src/ts/src/public/time-entry/dto.ts
@@ -1,0 +1,16 @@
+export interface TimeEntryDto {
+  id: string;
+  description: string;
+  start: Date;
+  end: Date;
+  deleted: boolean;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
+export interface TimeEntryUpdateValues {
+  start?: Date;
+  end?: Date;
+  description?: string;
+  deleted?: boolean;
+}